Move BoundingClientRectExtensions to MudBlazor.Extensions
